+++ modules/protocol/unrealircd.cpp
@@ -351,7 +351,7 @@ private:
 		if (Servers::Capab.count("ESVID") > 0 && !na->nc->HasExt("UNCONFIRMED"))
 			IRCD->SendMode(Config->GetClient("NickServ"), u, "+d %s", na->nc->display.c_str());
 		else
-			IRCD->SendMode(Config->GetClient("NickServ"), u, "+d %d", u->signon);
+			IRCD->SendMode(Config->GetClient("NickServ"), u, "+d %ld", (unsigned long)u->signon);
 	}
 
 	void SendLogout(User *u) override
